✅ test(distances,astro): isolate the optional-dependency tests
`pytest.importorskip("unxts.parametric")` sat inside six test bodies in three
modules that are otherwise about the core types. Nothing marked those modules
as touching an optional dependency, and a missing install showed up as
scattered individual skips.

One module per package now, gated once at import:

    if not OptDeps.UNXTS_PARAMETRIC.installed:
        pytest.skip("unxts.parametric is not installed", allow_module_level=True)

    from unxts.parametric import PQ

The skip has to precede the import -- it would otherwise fail at collection --
which is also why this cannot be a `pytestmark`. `PQ` is then imported plainly
at the top, so the tests read like ordinary tests instead of fetching the
symbol again in every body.

Follows the pattern `test_parametric_promotion` already used; that file is
renamed to `test_parametric` since it now holds the `from_` dispatch tests
alongside the promotion rules, and is the one place astro's optional-parametric
tests live.

Verified in an environment with `unxts.parametric` actually uninstalled: each
module reports a single skip naming the dependency, and 186 tests still pass.
The first attempt at that check was wrong -- assigning a sentinel string to the
enum member makes `.installed` return True, the opposite of the intent -- so
the package is genuinely removed instead.

4575 pass with it installed.
